// src/types/rotationNotifier.ts
	export interface TurfWarVsRule {
		name: "Turf War";
		rule: "TURF_WAR";
		id: "VnNSdWxlLTA=";
	}

	export interface TurfWarSetting extends BaseMatchSetting<TurfWarVsRule> {
		__typename: "RegularMatchSetting";
	}

	export interface TowerControlVsRule {
		name: "Tower Control";
		rule: "LOFT";
		id: "VnNSdWxlLTI=";
	}
	export interface ClamBlitzVsRule {
		name: "Clam Blitz";
		rule: "CLAM";
		id: "VnNSdWxlLTQ=";
	}
	export interface SplatZonesVsRule {
		name: "Splat Zones";
		rule: "AREA";
		id: "VnNSdWxlLTE=";
	}
	export interface RainmakerVsRule {
		name: "Rainmaker";
		rule: "GOAL";
		id: "VnNSdWxlLTM=";
	}

	export type RankedVsRule = TowerControlVsRule | ClamBlitzVsRule | SplatZonesVsRule | RainmakerVsRule;